pyproject.toml
[tool.poetry]
name = "nuros"
version = "0.1.0"
description = ""
authors = ["Ulan <ulan.sharipov@nu.edu.kz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.37"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
authlib = "^1.4.0"
pydantic-settings = "^2.7.1"
httpx = "^0.28.1"
itsdangerous = "^2.2.0"
google-auth = "^2.37.0"
pydantic = {extras = ["email"], version = "^2.10.5"}
asyncpg = "^0.30.0"
fernet = "^1.0.1"
redis = "^5.2.1"
python-jose = "^3.3.0"
sqladmin = "^0.20.1"
google-cloud-storage = "^3.0.0"
aioredis = "^2.0.1"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
